requirements.txt
langchain-google-genai
langchain-core
streamlit
Chroma
chromadb
google-generativeai
langchain
langchain-core
lark